requirements.txt
ffmpeg-python==0.2.0
picamera2==0.3.7
python-vlc==3.0.18121